engine/requirements.txt
playwright
google-generativeai
supabase
beautifulsoup4
python-dotenv
requests
lxml
schedule
fastapi
uvicorn
googlesearch-python

